requirements.txt
fastapi
uvicorn
pyautogui
pygetwindow
mss
pillow
psutil
requests
pywebview 